createAction convienence function and module renaming.
createAction function will return a factory function that will create a
new instance of the action with the correct dispatcher, a dispatch
method and any other custom methods that you may want to add. The
instance, once constructed, will have a dispatcher property so the
dispatcher for that action can be changed from the one defined as
default for that action.